pyproject.toml
[tool.poetry]
name = "mizu"
version = "0.1.0"
description = "Web chat interface for OpenAI language models similar to ChatGPT"
authors = ["Henry <hi@henry.moe>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.12"
uvicorn = "^0.34.2"
sqlalchemy = "^2.0.40"
databases = {extras = ["aiomysql"], version = "^0.9.0"}
python-dotenv = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
